refactor Methods to separate file
