+++ crates/topcoat-router/src/lib.rs
@@ -8,6 +8,7 @@ mod context;
 mod endpoint;
 mod error;
 mod layer;
+mod methods;
 mod module;
 mod page;
 mod path;
@@ -30,6 +31,7 @@ pub use context::*;
 pub(crate) use endpoint::Endpoint;
 pub use error::*;
 pub use layer::*;
+pub use methods::*;
 pub use module::*;
 pub use page::*;
 pub use path::*;
